% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipacsi.R
\name{ipacsi}
\alias{ipacsi}
\title{Perform calculation for Responents Satisfaction Index and Importance Performance Analysis}
\usage{
ipacsi(
  files,
  performance = "performance",
  importance = "importance",
  firstrow = TRUE,
  firstcol = TRUE
)
}
\arguments{
\item{files}{Input file from questionnaire. Only support Excel (\*.xls or \*.xlsx) file.}

\item{performance}{Character. Performance data sheet name in excel file}

\item{importance}{Character. Importance data sheet name in excel file}

\item{firstrow}{Logical. Is the attribute name at first row of file?}

\item{firstcol}{Logical. Is the respndent id at first column of file}
}
\value{
list
}
\description{
Measures Responents Satisfaction Index and Importance Performance Analysis from survey data.
}
\details{
\code{Result} contains percentage of suitability between performance and importance of a service or whatever you measure using quesionnaire, index of performance according to respondents and index for customer satisfactions. \code{Average} is overall average value of suitability, satisfaction and importance.
}
