% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{Launch Respondents Satisfaction Index Analysis Apps}
\usage{
launch(
  launch.browser = getOption("shiny.launch.browser", interactive()),
  port = NULL,
  host = getOption("shiny.host", "127.0.0.1"),
  display.mode = c("auto", "normal", "showcase")
)
}
\arguments{
\item{launch.browser}{If true, the system's default web browser will be launched automatically after the app is started. Defaults to true in interactive sessions only.}

\item{port}{The TCP port that the application should listen on. Defaults to choosing a random port.}

\item{host}{The IPv4 address that the application should listen on. Defaults to the shiny.host option, if set, or "127.0.0.1" if not.}

\item{display.mode}{The mode in which to display the example. Defaults to showcase, but may be set to normal to see the example without code or commentary.}
}
\description{
Shiny app will opened on your default browser
}
\examples{
\donttest{
launch()
}

}
