% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadplot.R
\name{quadplot}
\alias{quadplot}
\title{Plot for Customer Satisfaction Index and Importance Performance Analysis}
\usage{
quadplot(ipacsi, legend.pos = "right", ...)
}
\arguments{
\item{ipacsi}{List. Output from \code{ipacsi} function.}

\item{legend.pos}{Character. Available value is \code{"none"}, \code{"right"}, \code{"bottom"}.}

\item{...}{Argument to be passed to \code{ggplot2::theme()} function.}
}
\value{
plot
}
\description{
Display Customer Satisfaction Index and Importance Performance Analysis from survey data using ggplot2 output.
}
